requirements.txt
crewai==1.8.1
elasticsearch==9.2.1
fpdf2==2.8.5
langchain==1.2.6
langchain_community==0.4.1
numpy==2.4.1
pgvector==0.3.6
psycopg2==2.9.10
psycopg2_binary==2.9.11
python-dotenv==1.2.1
sentence_transformers==5.0.0
torch==2.7.1+cu118
urllib3==2.6.3
